Убрал ненужные файлы, объединил в один
