--- source/settings_bot.py
@@ -1,3 +1,0 @@
-# Токен бота
-
-TOKEN = '6478111175:AAHKn0haLwAn7dnCEIdDIkUxAhCSPuSmy64'
