--- source/utils_bot.py
@@ -1,6 +1,0 @@
-# Тут все функции связанные со всопомогательной работой бота
-from telegram import ReplyKeyboardMarkup, KeyboardButton
-
-
-def main_keyboard():
-    return ReplyKeyboardMarkup([['Начало работы']])
